refactor(number) : 파일 위치 이동
